pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rljb"
version = "0.1.0"
description = "Reinforcement-learning jailbreak experiments with LLMs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "llama-cpp-python>=0.2.0",
  "torch>=2.2.0",
  "numpy>=1.20.0",
  "transformers>=4.41.0",
]

[tool.setuptools.packages.find]
where = ["src"]
